+++ SmokeS/app/src/main/java/com/example/ayoubelyaghmouri/smokes/models/Longen.java
@@ -18,7 +18,7 @@ public class Longen extends Orgaan {
     @Override
     public void berekenPercentage(int nietGerookteSigaretten, int aantalMeldingen, DatabaseHelper db) {
 
-        final float BEGIN_PERCENTAGE = 60;
+        final double BEGIN_PERCENTAGE = 60;
         final int MAXIMALE_PERCENTAGE = 100;
 
         final double PROCENT_OMHOOG = 3;
